add UITests
